post

+++ app/Http/Controllers/TasksController.php
@@ -10,13 +10,23 @@ class TasksController extends Controller
     // getでTasks/にアクセスされた場合の「一覧表示処理」
     public function index()
     {
-        // タスク一覧を取得
-        $tasks = Task::all();
-
+        $data = [];
+        
+        if (\Auth::check()) { // 認証済みの場合
+        
+        // 認証済みユーザを取得
+        $user = \Auth::user();
+        
+        // ユーザの投稿の一覧を作成日時の降順で取得
+        $tasks = $user->tasks()->orderBy('created_at', 'desc')->paginate(10);
+            $data = [
+                'user' => $user,
+                'tasks' => $tasks,
+            ];
+        }
+        
         // タスク一覧ビューでそれを表示
-        return view('tasks.index', [
-            'tasks' => $tasks,
-        ]);
+        return view('tasks/index', $data);
     }
 
     // getでTasks/createにアクセスされた場合の「新規登録画面表示処理」
@@ -40,13 +50,28 @@ class TasksController extends Controller
         ]);
         
         // タスクを作成
-        $task = new Task;
-        $task->status = $request->status;
-        $task->content = $request->content;
-        $task->save();
+        $request->user()->tasks()->create([ 
+            'status' => $request->status,
+            'content' => $request->content,
+        ]);
 
-        // トップページへリダイレクトさせる
-        return redirect('/');
+        $data = [];
+        
+        if (\Auth::check()) { // 認証済みの場合
+        
+        // 認証済みユーザを取得
+        $user = \Auth::user();
+        
+        // ユーザの投稿の一覧を作成日時の降順で取得
+        $tasks = $user->tasks()->orderBy('created_at', 'desc')->paginate(10);
+            $data = [
+                'user' => $user,
+                'tasks' => $tasks,
+            ];
+        }
+        
+        // タスク一覧ビューでそれを表示
+        return view('tasks/index', $data);
     }
 
     // getでTasks/（任意のid）にアクセスされた場合の「取得表示処理」
@@ -54,11 +79,33 @@ class TasksController extends Controller
     {
         // idの値でタスクを検索して取得
         $task = Task::findOrFail($id);
-
-        // タスク詳細ビューでそれを表示
-        return view('tasks.show', [
-            'task' => $task,
-        ]);
+        
+        if (\Auth::id() === $task->user_id) {
+            // タスク詳細ビューでそれを表示
+            return view('tasks.show', [
+                'task' => $task,
+            ]);
+        }
+        
+        else if (\Auth::id() !== $task->user_id) {
+            $data = [];
+        
+            if (\Auth::check()) { // 認証済みの場合
+            
+            // 認証済みユーザを取得
+            $user = \Auth::user();
+            
+            // ユーザの投稿の一覧を作成日時の降順で取得
+            $tasks = $user->tasks()->orderBy('created_at', 'desc')->paginate(10);
+                $data = [
+                    'user' => $user,
+                    'tasks' => $tasks,
+                ];
+            }
+        
+            // タスク一覧ビューでそれを表示
+            return view('tasks/index', $data);
+        }
     }
 
     // getでTasks/（任意のid）/editにアクセスされた場合の「更新画面表示処理」
@@ -67,10 +114,32 @@ class TasksController extends Controller
         // idの値でタスクを検索して取得
         $task = Task::findOrFail($id);
 
-        // メッセージ編集ビューでそれを表示
-        return view('tasks.edit', [
-            'task' => $task,
-        ]);
+        if (\Auth::id() === $task->user_id) {
+            // メッセージ編集ビューでそれを表示
+            return view('tasks.edit', [
+                'task' => $task,
+            ]);
+        }
+        
+        else if (\Auth::id() !== $task->user_id) {
+            $data = [];
+        
+            if (\Auth::check()) { // 認証済みの場合
+            
+            // 認証済みユーザを取得
+            $user = \Auth::user();
+            
+            // ユーザの投稿の一覧を作成日時の降順で取得
+            $tasks = $user->tasks()->orderBy('created_at', 'desc')->paginate(10);
+                $data = [
+                    'user' => $user,
+                    'tasks' => $tasks,
+                ];
+            }
+        
+            // タスク一覧ビューでそれを表示
+            return view('tasks/index', $data);
+        }
     }
 
     // putまたはpatchでTasks/（任意のid）にアクセスされた場合の「更新処理」
@@ -81,17 +150,33 @@ class TasksController extends Controller
             'status' => 'required|max:10',
             'content' => 'required|max:255',
         ]);
-        
+            
         // idの値でタスクを検索して取得
         $task = Task::findOrFail($id);
-        
+            
         // タスクを更新
-        $task->status = $request->status;
-        $task->content = $request->content;
-        $task->save();
-
-        // トップページへリダイレクトさせる
-        return redirect('/');
+        $request->user()->tasks()->create([ 
+            'status' => $request->status,
+            'content' => $request->content,
+        ]);
+    
+        $data = [];
+            
+        if (\Auth::check()) { // 認証済みの場合
+            
+        // 認証済みユーザを取得
+        $user = \Auth::user();
+            
+        // ユーザの投稿の一覧を作成日時の降順で取得
+        $tasks = $user->tasks()->orderBy('created_at', 'desc')->paginate(10);
+            $data = [
+                'user' => $user,
+                'tasks' => $tasks,
+            ];
+        }
+            
+        // タスク一覧ビューでそれを表示
+        return view('tasks/index', $data);
     }
 
     // deleteでTasks/（任意のid）にアクセスされた場合の「削除処理」
@@ -99,10 +184,28 @@ class TasksController extends Controller
     {
         // idの値でタスクを検索して取得
         $task = Task::findOrFail($id);
-        // タスクを削除
-        $task->delete();
+        
+        // 認証済みユーザ（閲覧者）がその投稿の所有者である場合は投稿を削除
+        if (\Auth::id() === $task->user_id) {
+            $task->delete();
+        }
 
-        // トップページへリダイレクトさせる
-        return redirect('/');
+        $data = [];
+        
+        if (\Auth::check()) { // 認証済みの場合
+            
+        // 認証済みユーザを取得
+        $user = \Auth::user();
+            
+        // ユーザの投稿の一覧を作成日時の降順で取得
+        $tasks = $user->tasks()->orderBy('created_at', 'desc')->paginate(10);
+            $data = [
+                'user' => $user,
+                'tasks' => $tasks,
+            ];
+        }
+        
+        // タスク一覧ビューでそれを表示
+        return view('tasks/index', $data);
     }
 }
